refactor(macos): isolate native core support
